Bravo

+++ BE/Host.cs
@@ -6,7 +6,7 @@ namespace BE
 {
     public class Host
     {
-        public int HostKey { get; set; }
+        public long HostKey { get; set; }
         public string FirstName { get; set; }
         public string LastName { get; set; }
         public string PhoneNumber { get; set; }
